Place holder for submission success landing page
Now generated the qrcodes client side. This solves an issue with placing
the UUID qrcodes without revealing the public/private association. Now
it is all just present on the one volatile page.

TODO: final URLs still need to be implemented.

Also: Now all directories that should be public are exposed as static
dirs.