Fix dependencies of input fields
onChange only fires upon blur. The nextQuestion function is bound to onInput,
which fires on all input to the field. The dependency logic was bound to
onChange, thus the dependency info was only set after the next question
appeared.

Bad.

So, bind dependency/validation to onInput,a nd all is well.

Close #2