Initial stab at simplifying Timeline CSS; two questions converted
This throws away the floats and a decent amount of the HTML boilerplate.

However, only the side-by-side view works for now. The mobile, one-side view
does not yet work.

Also, some validation features were broken in the process.

Committing because there's too many pieces moving at once, and this is currently
in a fairly functional state.