Retire datalist; use checkboxes w/ field for appending
There were too many practical problems with using datalist. First off, support
on iOS is nonexistent (thanks to apple's renderer policy...) and IE's support is
horrifyingly buggy and doesn't fire onChange and onInput events (seriously).

On top of it, typing everything isn't nearly as user friendly (especially on
mobile), will lead to misspellings, and will likely suppress the number of items
users are willing to enter.

Checkboxes answer nearly all of this, with the exception that the software list
is /really/ long (183 ATM). So... that's going to have to be split up and/or
reduced as a next step.

And, as a bonus, there's two fewer requests for the client.